moving formatters into result builder
